Revert "TA#22109 Add columns target/suggested/profit to cost report (#133)"
This reverts commit 4e68d569f1ada117e274d916264f41acdc0a262e.
