+++ project_cost_report/i18n/fr.po
@@ -6,8 +6,8 @@ msgid ""
 msgstr ""
 "Project-Id-Version: Odoo Server 11.0+e\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2020-07-07 19:11+0000\n"
-"PO-Revision-Date: 2020-07-07 15:15-0400\n"
+"POT-Creation-Date: 2020-05-14 18:33+0000\n"
+"PO-Revision-Date: 2020-05-14 14:36-0400\n"
 "Last-Translator: <>\n"
 "Language-Team: \n"
 "Language: fr\n"
@@ -45,9 +45,9 @@ msgstr "Ligne analytique"
 
 #. module: project_cost_report
 #. openerp-web
-#: code:addons/project_cost_report/static/src/js/report_action.js:323
-#: code:addons/project_cost_report/static/src/js/report_action.js:368
-#: code:addons/project_cost_report/static/src/js/report_action.js:410
+#: code:addons/project_cost_report/static/src/js/report_action.js:305
+#: code:addons/project_cost_report/static/src/js/report_action.js:350
+#: code:addons/project_cost_report/static/src/js/report_action.js:392
 #, python-format
 msgid "Analytic Lines ({category})"
 msgstr "Lignes analytiques ({category})"
@@ -122,13 +122,6 @@ msgid "FOLD"
 msgstr "PLIER"
 
 #. module: project_cost_report
-#. openerp-web
-#: code:addons/project_cost_report/static/src/xml/templates.xml:21
-#, python-format
-msgid "HIDE SUMMARY"
-msgstr "CACHER SOMMAIRE"
-
-#. module: project_cost_report
 #: model:ir.model.fields,field_description:project_cost_report.field_project_cost_category_id
 #: model:ir.model.fields,field_description:project_cost_report.field_project_cost_report_id
 msgid "ID"
@@ -170,7 +163,6 @@ msgstr "Nom"
 #. module: project_cost_report
 #: model:ir.ui.view,arch_db:project_cost_report.cost_report_body__outsourcing_section
 #: model:project.cost.category,name:project_cost_report.cost_category_outsourcing
-#: selection:project.cost.category,section:0
 msgid "Outsourcing"
 msgstr "Sous-traitance"
 
@@ -189,14 +181,8 @@ msgstr ""
 #. module: project_cost_report
 #: model:ir.ui.view,arch_db:project_cost_report.cost_report_body__product_section
 #: model:project.cost.category,name:project_cost_report.cost_category_product
-#: selection:project.cost.category,section:0
 msgid "Products"
 msgstr "Articles"
-
-#. module: project_cost_report
-#: model:ir.ui.view,arch_db:project_cost_report.cost_report_body__costs
-msgid "Profit"
-msgstr "PROFIT"
 
 #. module: project_cost_report
 #: model:ir.actions.act_window,name:project_cost_report.action_project_cost_category
@@ -230,36 +216,14 @@ msgid "Reference"
 msgstr "Référence"
 
 #. module: project_cost_report
-#. openerp-web
-#: code:addons/project_cost_report/static/src/xml/templates.xml:17
-#, python-format
-msgid "SHOW SUMMARY"
-msgstr "AFFICHER SOMMAIRE"
-
-#. module: project_cost_report
 #: model:ir.ui.view,arch_db:project_cost_report.project_cost_category_search
 msgid "Search Project Cost Categories"
 msgstr "Chercher des catégories de rapport de coût"
 
 #. module: project_cost_report
-#: model:ir.model.fields,field_description:project_cost_report.field_project_cost_category_section
-msgid "Section"
-msgstr "Poste"
-
-#. module: project_cost_report
 #: model:ir.model.fields,field_description:project_cost_report.field_project_cost_category_sequence
 msgid "Sequence"
 msgstr "Séquence"
-
-#. module: project_cost_report
-#: selection:project.cost.category,section:0
-msgid "Shop Supply"
-msgstr "Matériel"
-
-#. module: project_cost_report
-#: model:ir.ui.view,arch_db:project_cost_report.cost_report_body__costs
-msgid "Suggested"
-msgstr "Suggéré"
 
 #. module: project_cost_report
 #: model:ir.ui.view,arch_db:project_cost_report.cost_report_body__costs
@@ -268,34 +232,18 @@ msgid "Supplier"
 msgstr "Fournisseur"
 
 #. module: project_cost_report
-#: model:ir.ui.view,arch_db:project_cost_report.cost_report_body__costs
-msgid "Target"
-msgstr "Objectif"
-
-#. module: project_cost_report
-#: model:ir.model.fields,field_description:project_cost_report.field_project_cost_category_target_hourly_rate
-msgid "Target Hourly Rate"
-msgstr "Objectif taux horaire"
-
-#. module: project_cost_report
-#: model:ir.model.fields,field_description:project_cost_report.field_project_cost_category_target_margin
-msgid "Target Margin"
-msgstr "Objectif marge"
-
-#. module: project_cost_report
 #: model:ir.model,name:project_cost_report.model_task_type
 msgid "Task Type"
 msgstr ""
 
 #. module: project_cost_report
-#: code:addons/project_cost_report/report/report.py:140
+#: code:addons/project_cost_report/report.py:88
 #, python-format
 msgid "The cost report was triggered without a project ID in context."
 msgstr "Le rapport de coûts a été déclenché sans passer un ID de projet en contexte."
 
 #. module: project_cost_report
 #: model:ir.ui.view,arch_db:project_cost_report.cost_report_body__time_section
-#: selection:project.cost.category,section:0
 msgid "Time"
 msgstr "Temps"
