+++ project_cost_report/models/project_cost_category.py
@@ -11,14 +11,5 @@ class ProjectCostCategory(models.Model):
     _order = "sequence"
 
     sequence = fields.Integer()
-    name = fields.Char(translate=True, required=True)
+    name = fields.Char(translate=True)
     active = fields.Boolean(default=True)
-
-    section = fields.Selection(
-        [("products", "Products"), ("time", "Time"), ("outsourcing", "Outsourcing")],
-        required=True,
-        default="products",
-    )
-
-    target_margin = fields.Float()
-    target_hourly_rate = fields.Float()
