--- project_cost_report/report/__init__.py
@@ -1,4 +1,0 @@
-# © 2020 - today Numigi (tm) and all its contributors (https://bit.ly/numigiens)
-# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
-
-from . import report
