+++ project_cost_report/static/src/less/cost_report.less
@@ -77,10 +77,12 @@
 }
 
 /**
- * Add borders around the totals
+ * Add borders around the total cost amount
  */
-td.o_project_cost_report__total {
-    border: 1px black solid;
-    padding: 8px;
-    width: 200px;
+.o_project_cost_report__total_cost {
+    td:last-child {
+        border: 1px black solid;
+        padding: 8px;
+        width: 200px;
+    }
 }
